Rename job.py to testing_job.py.
- Remove unused credentials option in testing job init methods.
- Rename test_image method to process_job a more generic name.
