+++ mash/services/testing/ec2_job.py
@@ -18,7 +18,7 @@
 
 import random
 
-from mash.services.testing.job import TestingJob
+from mash.services.testing.testing_job import TestingJob
 
 instance_types = [
     'c5d.large',
@@ -40,7 +40,7 @@ class EC2TestingJob(TestingJob):
 
     def __init__(
         self, id, last_service, cloud, ssh_private_key_file, test_regions,
-        tests, utctime, job_file=None, credentials=None, description=None,
+        tests, utctime, job_file=None, description=None,
         distro='sles', instance_type=None, ipa_timeout=None,
         ssh_user='ec2-user'
     ):
