+++ mash/services/testing/service.py
@@ -139,7 +139,7 @@ class TestingService(PipelineService):
         Test image with IPA based on job id.
         """
         job = self.jobs[job_id]
-        job.test_image()
+        job.process_job()
 
     def _get_listener_msg_args(self):
         """
